plenty of new additions
the registration bar finally leads to separate registration pages both for login and register as a bonus logout actually asks you if you are sure you want to log out. the tutorial browser finally works and can store tutorials. the tutorial button now uses both the username and tutorial title to generate a unique url for when you click on it.I will use this when a tutorial can lead to a page. After way way way tool long and lots of google searches and inspiration(stealing) from templates the filters for the tutorial browser actually work aside from the text search field which i start to doubt i need but i could implement it in the future if i have time. Right sidebar is still only a visual component.a css file for the register component has been aded and i think this might be it for now.