profile page,infinite loop cause discovered and fixed, and some better user registration functionlality
UserCnotext has been update to allow to save the user after a page refresh by using localStorage. A bug showed up that led to the user being both registered but not exist. This happened because after stopping the softuni practice server the user would no longer exist but the localstorage would still have the information of the user being saved. I fixed this by adding a check in UserContext on reload to check if the the user acess-token exists. If not the user is set to null and the local storage item that corresponds to the no longer existing user is removed. That way when turning off the server the user is no longer in a weird limbo state between logged in but not really while still allowing the user to stay logged in on page refresh. The search text bar filter in the tutorila browser finally filters the tutorials by their titles while working together with the other filters. The infinite re-render loop was finally removed.... at least i think so. And  the user Profile page was finally added  which was relatively straight forward and simple now that the registration actually works. If i am not mistaken all thta is left is to add the ability to  create tutorial .