the ability to create a tutorial has been added
No more need for built in tutorials as now there is a way for the user to create one. As of now everyone including the unregistered users can make tutorials(subject to change) the tutorialCreation page now stores step information iside an array meaning that the add steps button can add more steps to the array. The ability to add an image has been added (shamelessly stolen from the internet). TutorilaContext.jsx has been made obsolete and will probably be removed n the future but i will keep it for now just in case something doesnt work as i expect. The tutorials are sent directly to the localStorage as an object: {
  "title": "Character Modeling",
  "category": "Modeling",
  "platform": "Blender",
  "coverImage": "data:image/png;base64,iVBORw0KGgoAAAANS...",
  "steps": [
    { "title": "Step 1", "description": "Do this", "image": "data:image/png;base64,..." },
    ...
  ],
  "creator": "user@example.com",
  "date": "2025-12-05T12:34:56.789Z"
}
as of now it seems to work.
Remaining stuff: edit, delete, actyal place to view the tutorials as a user(TutorialViewPage or something),and make the tutorial ceration button only available to registered users