even more stuff added like the ability to actually read a tutorial as a user and route guards
A very simple route guard has been added Unregistered users can no longer make tutorials or access places they shouldnt but in case they try they will be redirected to the login page. images uploaded in the tutorial now get compressed so that they take up less storage and the id of the tutorial has been changed to be the date of creation as the previous method of getting the user email didnt always work. A page for the tutorials themselves has been added that lets the users read the tutorial. Some bugs and inconsistencies with the search filters have been fixed. By some miracle TutorialSelection was still working although the import was referring to it as TutorialBrowser as it was called that before. Now it is imported with the correct name. Inside the Tutorial selection component useEffect has been added to get data on the spot in order to fix a bug where on refresh the pre-selected category would update before the tutorial got loaded so you would get no results from the pre-selected category until manually selecting it. The tutorial view as i call it now has its own css file tied to it. Edit and Delete functionality is yet to be implemented