delete and edit are added + some minor changes to the routes in app.jsx
Two new buttons have been added that are only acessable to the creator of a given tutorial Edit and Delete. Delete's functionality is directly inside the TutorialView and simply removes the tutorial from local storage by using its id. The Edit button lead to a new page that is a copy of the creation page but it gets filled with the information from the tutorial you are editing. The  new things inside it are the update step function and the save changes function(mostly copied from publish  tutorial function from the creation page) which simply overwrites the old version of the tutorial. Because the edited version is a direct copy of the original it still keeps all the meta data: creator, date of creation , id(just the exact time of creation in miliseconds by using Date.now()). in Addition there are now checks if the user is loged in to prevent a loged user from accessing register and login. in case the user tries they will be redirected to the tutorials page.
This should cover the assignment requirements.