minor quality of life changes and checks to ensure propper functionality
the tutorial selection now informs the user of the lack of tutorial instead of just showing nothing. The user is now using Session storage for persistance instead of local storage in case it doesnt fit the assignment requirements. Tutorial page no longer gives an annoying popup informing that tutorials dont exist every time the user enters the page. instead it gives a notification in the console for the developer. This does not in any way break or prevent functionality of the app.